+++ exid/uuid.go
@@ -14,12 +14,6 @@
 package exid
 
 import (
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/ergoapi/util/exhash"
-
 	"github.com/google/uuid"
 )
 
@@ -33,17 +27,4 @@ func GenUUID() string {
 func CheckUUID(uid string) bool {
 	_, err := uuid.Parse(uid)
 	return err != nil
-}
-
-// Deprecated: use HashUID instead
-func GenUID(username string) string {
-	return exhash.MD5(username + fmt.Sprint(time.Now().UnixNano()))
-}
-
-// HashUID 生成新的uid
-func HashUID(username string, prefix string) string {
-	if len(prefix) > 0 && !strings.HasSuffix(prefix, "-") {
-		prefix = fmt.Sprintf("%s-", prefix)
-	}
-	return fmt.Sprintf("%s%s", prefix, exhash.MD5(username+fmt.Sprint(time.Now().UnixNano())))
 }